pyproject.toml
[project]
name = "virtual_triage_system"
version = "0.3.3" # 3rd week of sprint 3
description = "UMBC CSMC 447 Fall 2025 Team 5 Project"
dependencies = [ "flask",
                 "sqlalchemy",
                 "markdown-it-py",
                 "Whoosh",
                 "openai",
                 "Flask-Bcrypt",
                 "xdg-base-dirs",
                 "irc",
                 "psycopg2-binary" ]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["vts*"]
